# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Massage Booking API",
    "version": "1.0.0",
    "description": "API documentation for massage booking application"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "Massage": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a1f2b3c4d5e6f7a8b9c0d2"
          },
          "name": {
            "type": "string",
            "example": "Serene Touch Massage"
          },
          "address": {
            "type": "string",
            "example": "123 Main Street"
          },
          "district": {
            "type": "string",
            "example": "Pathum Wan"
          },
          "province": {
            "type": "string",
            "example": "Bangkok"
          },
          "postalcode": {
            "type": "string",
            "example": "10330"
          },
          "tel": {
            "type": "string",
            "example": "02-123-4567"
          },
          "pictures": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://placehold.co/600x400"
            ]
          },
          "price": {
            "type": "number",
            "example": 500
          },
          "ratingSum": {
            "type": "number",
            "example": 45
          },
          "userRatingCount": {
            "type": "number",
            "example": 10
          },
          "averageRating": {
            "type": "number",
            "example": 4.5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MassageInput": {
        "type": "object",
        "required": [
          "name",
          "address",
          "district",
          "province",
          "postalcode",
          "tel",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Serene Touch Massage"
          },
          "address": {
            "type": "string",
            "example": "123 Main Street"
          },
          "district": {
            "type": "string",
            "example": "Pathum Wan"
          },
          "province": {
            "type": "string",
            "example": "Bangkok"
          },
          "postalcode": {
            "type": "string",
            "example": "10330"
          },
          "tel": {
            "type": "string",
            "example": "02-123-4567"
          },
          "pictures": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://placehold.co/600x400"
            ]
          },
          "price": {
            "type": "number",
            "example": 500
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/massages/{id}": {
      "get": {
        "summary": "Get a single massage shop",
        "tags": [
          "Massages"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Massage shop ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Massage shop details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Massage"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Massage shop not found"
          }
        }
      },
      "put": {
        "summary": "Update a massage shop (Admin only)",
        "tags": [
          "Massages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Massage shop ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MassageInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Massage shop updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Massage"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - Not logged in"
          },
          "403": {
            "description": "Forbidden - Not an admin"
          },
          "404": {
            "description": "Massage shop not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a massage shop (Admin only)",
        "tags": [
          "Massages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Massage shop ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Massage shop deleted (associated reservations are also removed)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not logged in"
          },
          "403": {
            "description": "Forbidden - Not an admin"
          },
          "404": {
            "description": "Massage shop not found"
          }
        }
      }
    },
...
  }
}